requirements.txt
anyascii==0.3.3
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.12.3
certifi==2026.2.25
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-taggit==6.1.0
django-treebeard==4.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
draftjs_exporter==5.2.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
l18n==2021.3
laces==0.1.2
openpyxl==3.1.5
pillow==11.0.0
pillow_heif==0.22.0
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
telepath==0.3.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wagtail==6.3.1
Willow==1.11.0
